[#MODX-1734] Add context parameter to the getLoginUserName() API function
Commit emulates getLoginUserID()'s acceptance of a $context parameter. If $context
is not supplied, then behavior of the function is unchanged. This change allows,
for example, the username of a manager user to be retrieved through the API from the
front end if that user is also logged in as a web user.