[#MODX-1035] XSS vuln. in installer collation processor
In the event that the collation processor was unable to connect to the database,
a POST variable was directly being written to an html select element representing
the db collation for the install. This commit sanitizes the value using
htmlentities. Further sanitization of POST variables are done in the database test
processor. Further testing suggested.