[#508] fix SQL error in module dependencies
In some configurations, var_export was causing a php array to be injected
into the SQL that updates the active_users table. var_export is only
needed when the variable in question is null, so the code has been
modified to only use var_export when the value is null, otherwise the
value, which has been verified as an integer, is used directly.