[#421] new overwrite flag for Chunks in installer
In installer Chunk phpdoc, @internal @overwrite false will flag the
installer to not overwrite a Chunk if one exists with the same name.
This allows ManagerManager, for example, to continue to use a
rules Chunk named mm_demo_rules and not overwrite a user's customized
version. So that the Chunk still gets installed, it is renamed, with
the MODx version number appended to it's name.