[#2867] update to handling of demo elements
All file-based installer elements are now flagged as being part of the demo
content. This may be pared down a little, but the idea is that if the user
selects that they want the demo site installed, we need to ensure that the
user cannot unselect elements that the demo actually depends on.

Un-checking the demo site on the Options page now leaves the elements'
checkboxes checked, simply enabling them for user input again.