Better modx_sanitize_gpc - 2nd try
In the old version, MODX tags separated in two superglobal array members were not sanitized. But if they are shown in the right order, the tags are evaluated by the parser. Extended/modified the code from the japanese community.
Tags are converted into hexadecimal entities now. Otherwise a [<b>[ tag is not sanitized but after a strip_tags it is a valid MODX tag.