Fixed CodeMirror syntax HTML entity highlighting
Updated the custom MODx CodeMirror mode to better highlight HTML entities. As it was written before, it treated anything between an ampersand (&) and an equal sign (=) as a snippet property. Unfortunately, this meant that any HTML that contained entities like so:
<p>Hello &amp; World!</p>
would be highlighted blue all the way to the end of the line.

I updated the custom mode with a Regular Expression to correct the problem