Maybe missing permission check
I'm not sure about the purpose of the field "Empty the site's cache" under "Content management" in the user-roles form. I would think it allows or disallows the user to empty the site's cache by clicking on the menu item "Site -> Clear cache". That doesn't work because there is no permisson check in refresh_site.dynamic.php.
If there is a different purpose, feel free to enlighten me. :-)