Workaround for deprecated preg_replace() with /e modifier
Hi there,

I only came across this error, because I wrote a snippet with a function inside, which was not enclosed in `ìf(!function_exists())` and that resulted in a blank page. Otherwise the error "preg_replace(): The /e modifier is deprecated, use preg_replace_callback instead" only shows up in the system log, but the page is rendered as usual.

My server runs on PHP 5.5.29. It works for me, but I didn't test it with all configurations. I refinded my original code with the help of http://pastebin.com/2fAGjfkW in [issue 438] (https://github.com/modxcms/evolution/issues/438) which I found after finishing my original workaround...