Concept on premissions to edit res-type for only mgrRole=1 or rsc-creators?
I need my editors to be able to change ressource-type. They only be allowed to do so on ressources created by their own!
As a temporary fix I added mgrRole 2 to the condition.

Can anybody explain the concept behind that restriction? 
Wouldn't it be worth a setting in the security›role section?