New API - $modx->stripslashes()
echo stripslashes($_GET['v']);
Such code has vulnerability.
echo $modx->stripslashes($_GET['v']);
If we write like this, no problem will occur.
Originally, stripslashes is not a very dangerous function.
However, if we write code to read GPC's value in MODX, problems are
likely to occur, so we need to be careful.
And the stripslashes function should be almost unnecessary in PHP 5.4
and later since magic quotes are obsolete.