[I] changing output filename for easier sorting
Hi there,

maybe it's just me, but when developing a site I sometimes need to delete the cached images from phpthumb. But when there are several images in different sizes of on image, it's awful to select them from a list sorted by filename. So instead of the current output filename `1200x1200-filename-123.jpg` I propose to generate it like this `filename-1200x1200-1234.jpg` where `1234` is the hash. Then all generated files are directly one below the other.