[I] Link snippet name to arbitrary function
MODX links snippet name and PHP code.
In addition to this, in the case of a PHP code string that can be
recognized as a PHP function name, the snippet name and the PHP function
name are linked.

ex)
[php]
$modx->addSnippet('MySnippet','myFunction');
function myFunction($param) {
return $param['name'];
}
[html]
[[#MySnippet?name=test]]

It is similar to "add_action" in WordPress.