Update snippet.docinfo.php (#806)
* Update snippet.docinfo.php

Imagine you have a catalog of products, which are graded by categories. Categories are represented by parental documents. And you need to show the name of the category (or other parent's TV's) on the product's page. You can't use &docid of the DocInfo snippet in the product's template, because ID is different for different categories, but the template is only one.

With the new parameter &templid you can get field's value of the document, which has the template's ID set to the &templid. Script finds that document among all parents of &docid - beginning from document with &docid and upper and upper.

books (catalog - id: 1, template: 1)
--romance (category - id: 2, template: 2)
----book1 (product - template: 4)
----book2 (product - template: 4)
--fantasy (category - id: 3, template: 2)
----for children (subcategory - id: 5, template: 3)
------book5 (product - template: 4)
----book3 (product - template: 4)
----book4 (product - template: 4)
--thriller (category - id: 4, template: 2)

In the template with ID 4 (for book) we can get `pagetitle` of categories:

[[DocInfo? &templid=`2`]]
it will return "romance" for the book1 and book2, "fantasy" for the book3, book4 and book5! (subcategory "for children" hasn't template's ID = 2)

To get any TV of category:
[[DocInfo? &templid=`2` &field=`tvname`]]

The same in the chunk with placeholders (for example, DocLister's template):
[[DocInfo? &docid=`[+id+]` &templid=`2` &field=`tvname`]]
Note you don't need to add the field `tvname` to the &tvList in the DL call for using in the DocInfo only.

* Update snippet.docinfo.php

1. The &limit parameter was added
2. Fix false value of $modx->getDocument

* Update snippet.docinfo.php

* Update snippet.docinfo.php